+++ CHANGELOG.md
@@ -7,4 +7,4 @@
 - Add support for Windows CI and testing with WSL (#184)
 - Add StackExchange.Redis compatible pub/sub API (#202)
 - Add transparent compression support with Zstd and LZ4 backends (#213)
-- Add Valkey Search command support: FT.CREATE, FT.DROPINDEX, FT._LIST, FT.SEARCH, FT.AGGREGATE, FT.INFO, FT.ALIASADD, FT.ALIASDEL, FT.ALIASUPDATE, FT._ALIASLIST (#225)
+- Add Valkey Search command support (#225)